Implement sets.
Sets are a python builtin type that may only contain one instance
of a value. This set type achieves this by using the keys of a dict
as the storage for its elements. Some of the standard set functions
are implemented here, but by no means all.

This commit also adds a collection of tests for the functionality
added to date, some of which depends upon the commits before this
one in this tree. Notably the length of a dictionary, and the
comparison of list objects that are the same item. The tests for
those are numbered after these as that was the order of issue
discovery and fixing.