Big Merge of Bob's and Brad's Changes The way time was handled was greatly changed, so that speed changes took effect.  I also changed the actual animation render() code to handle this, to add more functions (for instance, to let you clear the canvas to do animation beyond animating the turtle) and I changed the loop to use the Javascript setTimeout function, so it doesn't seize control of the browser (each step in the loop re-executes render() with setTimeout, instead of an actual loop).
Brads fixes to the color variables were incorporated previously in a different way, because I'd expanded those functions to take parameters in two fashions (as a tuple of rgb components or as 3 separate parameters, as well as a string value).

To support animation, I added delay, setDelay and clear functions.