Make quit() and exit() work Issue #36
This is Issue 115 in the Google Code repo

I did this as a language change making quit() and exit() flow
statements.

I added a unit test which nicely tests functionality and passes
but the parser and symbol table unit tests are unhappy.   Either
this is a problem in my version of Python or I missed some detail
on a step I needed to take other than regentests after adding a
new unit test.  Thanks in advance for any help/explanation here.