Roll back the changes that added line numbers to NameError
There is an easier and more consistent way to add this to the runit wrapper function used by skulpt applications.  Maybe there is a way to add this same level of consistency to the skulpt core.  But for now, it seems safer to roll back this change and grab the line numbers in the wrapper.