Restore Sk.tg.fadeOnExit
This works as long as the turtle module is used in the context of skulpt.  If you try to use the turtle graphics implementation as a standalone javascript turtle then you will get this error.  But I'm willing to live with that since the turtle module is supposed to be used in conjunction with Skulpt.