Fixes ondrag implementation
Previously using ondrag required also calling onclick. This is not how the Python
version works so this PR fixes ondrag such that it can be used on its own.

Essentially the MouseManager is always listening for events now and on a mousedown/mouseup
it marks any mousemove (ondrag) managers as elligible for dragging.

This PR also includes a couple of minor style fixes to get jshint happy