Fast path for function tp$call()
This saves about 1% on @rixner's sample program.

According to Chrome's profiler, this change takes func.tp$call()'s
self-time down from 6.4% to 5.8%. (And another 4.0% on
method.tp$call()!). I'd say this is still a significant optimisation
target - the trick is to hit it without causing subtle correctness
issues (which have plagued this code in the past).