Only override tp$getattr() when a class defines __getattribute__
This saves 10.3% runtime on @rixner's sample app.

Previously, we were doing a typeLookup on every tp$getattr() on
a user-defined class, which is inefficient - the class either
defines __getattribute__ or it doesn't! So now we only generate
a custom tp$getattr() if we need it - the rest of the time, we
inherit from `object`. (Sadly, we can't actually use the standard
dunderToSkulpt[] mapping for this, as tp$getattr() has a slightly
different API to __getattribute__: in JS we signal "not found"
with undefined, but in Python it's an AttributeError, so we need
custom code to translate.)