Add a fast path for looking up strings in dict buckets.
This saves 13.6% runtime on @rixner's sample program.

We spend a lot of time looking up strings in dictionaries (we do it
on every standard attribute access, for starters), and invoking the
full richCompareBool() machinery each time was slowing us down.

So now we special-case looking up strings within a bucket, and do a
JS equality check.