Prototype "fast-call" function calling convention
This saves 2.4% of runtime on @rixner's example app.

Previously, every function call had to bounce through func.tp$call,
which assembled arguments based on the function's (formal) arguments
and then call()ed the underlying function.

This change allows a function to request "fast call", where func_code
itself *is* the tp$call method of the func object, and calls a method
of func to resolve its actual->formal arguments.

This opens the door to specialising the argument-resolution process
for functions with simple arguments.