Change breakpoint in while statement for debugging
Currently the breakpoint in a while loop is set to be after the test.
This means that when implementing a line by line debugger, the final
test that causes the while loop to terminate is not shown. By changing
the position of the debug block, the final test will be shown.